+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added support for workspace symbols to resolve symbols across the whole workspace. In VSCode, you can open this using `Ctrl + T`
+
 ## [1.19.2]
 
 ### Fixed